Add server API to access database and template engine
Add the web server API to access the database resources: Customer, Seller
and the Products.

Also added Embedded JavaScript templating (ejs) to modify the html page
adding the data from the database before sending it to the client. This
also changed the project structure, now the html files are located under
the views folder.