Implement methods to login & register users by querying db
Filled the body of the methods to register a seller or a customer and to
perform login of sellers or customers. This also implies the SQL queries
needed to retrieve the information from the database.

The methods were tested with Postman by sending a JSON. Probably the final
version would not use JSON but as application/x-www-form-urlencoded,
but it should still work correctly then although tested is needed.

Also added some more config code to use the ejs, but not code using it
yet.