Add support for creating and updating products
Add support to create new products in the database or to update the
ones already present. This features were tested using the test.html file
in the views folder which contains a html form.

An important milestone was the addition of Multer, a Express middleware
to handle multipart/form-data messages, the common way to send images.
This enables to upload the product image at creation of a new product or
when updating an already existing one.