Fix search icon bug and fix server paths problem
The search icon id was mistakenly assigned to the shopping cart icon in
the html file and viceversa. Reverse the ids so that they refer to the
correct icon.

The toggleElement function had a bug when replacing the w3-hide class.
It would only work if there were more class names in the element, as it
looked for an empty space before the class name, but in the search form
the only class was w3-hide and the error appeared.

The test in the development system use as source path the folder where
the html index file is present, but in the server this does not work.
To add the relative path of the resources and not having to change the
file structure in the server (so it can mirror the structure in development)
I add some express.static() mappings.