Add seller page and product page
Now sellers are able to see what products they are selling in our store
and they can add new products.

There is a product page that can be access at the moment only from the
sellers, but in the future also from any other page where products are
displayed. In this new page you can see the product characteristics and
as a seller you can modify the products.