First attempt at adding product to shopcart, update database
Need to retest the process of adding products to shopping cart.

Changed the database to have a new field for shopping carts indicating
if it was bought already. I may need also to add a quantity to the order
table to know how many products of a type the customer wants to buy.