Modify database after sprint review to fix issues found
During Sprint review there was some confusion with the ORDER table, as
its name may convey more meaning than intended. Renamed table to
PRODUCTS IN CART to avoid confusion. Also added a Quantity column to
specify the quantity of products chosen, enabling to have more than one
of each type in a SHOPPING CART.

Add a date to SHOPPING CART to represent the last time changed, or if it
has been bought the time of paying.