Add functionality to retrieve customer shopping carts
Add GET method to server where with a customerID we can retrieve the
customer shopping carts. For this purpose we add two new SQL queries to
retrieve all the shopping carts of a customer and to retrieve all the
products of a shopping cart.

This feature is available in web page through the shopping cart icon in
the customers GUI, so a new link was made from the icon to the server path.