Fix bug in SQL query in Quantity property and Price shown
Query will join two tables with the same column name for Quantity, hence
only the second value stayed which was incorrect. Limit query to only
retrieve the quantity being selected in that shopping cart and not the
total quantity of that product.

When rendering page the values used to calculate shopping cart price where
the actual values of the product and not the ones stored when the product
was added to the shopping cart as intended.