Add functionality to remove product from shopping cart
Add a button in each product row to delete that product. It uses AJAX so
we were not limited to the GET and POST HTTP methods and we could use
DELETE.

Also added a visual warning in the product rows when the quantity
requested is superior to the quantity available of that product.