Add functionality to set quantity when buying product
Add new functionality to specify the quantity of a product that wants to
be added to that customer shopping cart.

Found typo bug when checking the "length" property of a product. It was
mispelled as "lenght", so many server path had this issue and now their
behaviour changed, but most of the functionality was re-tested and it
should work as before.