Add functionality to create comments and correct visualization
Add the functionality to the POST button under the comment text block
used as input.

Fix how the comments are displayed by not escaping the Html when inserted
with ejs, such that the line breaks(<br>) are correctly displayed.
Then changed preset comments from populate database script to test this
feature.